Move implementation of implicit_function() into .cc
